+++ example/src/lib.rs
@@ -4,22 +4,23 @@
 //! `#![feature(proc_macro_hygiene)]`, so the following diagram is rendered by
 //! [`svgbobdoc::transform_mdstr!`] instead.)
 //!
-#![doc = svgbobdoc::transform_mdstr!(
+#![doc = transform_mdstr!(
 //! ```svgbob,
 //!  .--------------------.
 //!  | Diagrams here      |
 //!  `--------------------'
 //! ```
 )]
-#![doc = svgbobdoc::transform_mdstr!("
+#![doc = transform_mdstr!("
 ```svgbob,
  .--------------------.
  | Diagrams here      |
  `--------------------'
 ```
 ")]
+use svgbobdoc::transform_mdstr;
 
-#[svgbobdoc::transform]
+#[doc = transform_mdstr!(
 /// Some module.
 ///
 /// ```svgbob,
@@ -27,13 +28,18 @@
 ///  | Diagrams here |
 ///  `---------------'
 /// ```
+)]
 pub mod module {
+    #![doc = transform_mdstr!(
     //! ```svgbob,
     //! hoge
     //! ```
+    )]
+
+    use svgbobdoc::transform_mdstr;
 }
 
-#[cfg_attr(doc, svgbobdoc::transform)]
+#[doc = transform_mdstr!(
 /// Some function.
 ///
 /// ```svgbob,
@@ -41,9 +47,10 @@ pub mod module {
 ///  | Diagrams here      |
 ///  `--------------------'
 /// ```
+)]
 pub fn test_function() {}
 
-#[svgbobdoc::transform]
+#[doc = transform_mdstr!(
 /// Some structure.
 ///
 /// ```svgbob,
@@ -51,21 +58,19 @@ pub fn test_function() {}
 ///  | Diagrams here      |
 ///  `--------------------'
 /// ```
+)]
 pub struct TestStruct {
-    /// Fields [can't have] attribute macros, so the struct's `#[transform]`
-    /// handles the fields as well.
-    ///
-    /// [can't have]: https://github.com/rust-lang/rust/issues/53012
-    ///
+    #[doc = transform_mdstr!(
     /// ```svgbob,
     ///  .--------------------.
     ///  | Diagrams here      |
     ///  `--------------------'
     /// ```
+    )]
     pub field1: u32,
 }
 
-#[svgbobdoc::transform]
+#[doc = transform_mdstr!(
 /**
  * Some impl.
  *
@@ -75,12 +80,14 @@ pub struct TestStruct {
  *  `--------------------'
  * ```
  */
+)]
 impl TestStruct {
-    #[svgbobdoc::transform]
+    #[doc = transform_mdstr!(
     /// Some method.
     ///
     /// ```svgbob,
     /// hoge
     /// ```
+    )]
     pub fn test_method() {}
 }