+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Breaking** Updated `svgbob` to 0.5.
 - **Breaking** Code blocks indented by more than three spaces are now processed.
 - **Breaking** The macros are no-op by default. The `enable` Cargo feature must be enabled for the transformation to take place.
-- Added `transform_mdstr!`.
+- Added `svgbobdoc::transform_mdstr!`.
+- `#[svgbobdoc::transform]` is now deprecated.
 
 ## [0.2.3] - 2020-10-22